// src/utils/models/colorTypeEnum/color-type.enum.ts
export enum ColorType {
    Default = '/public/images/Texture/Default.png',
    Level1 = '/public/images/Texture/Level1.png',
    Level2 = '/public/images/Texture/Level2.png',
    Level3 = '/public/images/Texture/Level3.png',
    Level4 = '/public/images/Texture/Level4.png',
    Level5 = '/public/images/Texture/Level5.png',
    Level6 = '/public/images/Texture/Level6.png',
    Level7 = '/public/images/Texture/Level7.png',
    Level8 = '/public/images/Texture/Level8.png',
    Level9 = '/public/images/Texture/Level9.png',
    Level10 = '/public/images/Texture/Level10.png',
}